// external/@wdk/account-abstraction-ton/packages/ui-stargate-v2-sdk/src/typechain/StargatePool.ts
export type MessagingFeeStruct = {
  nativeFee: PromiseOrValue<BigNumberish>;
  lzTokenFee: PromiseOrValue<BigNumberish>;
};